Fixed issue regarding DOM querying in react.
